Move Fragmenter into preference_comparisons
Seems unlikely to be used elsewhere and doesn't
make much sense to have inside imitation.data.
